python/input-to-s3/requirements.txt
boto3
scramjet-framework-py
